# By explon (4) and kk (1)
# Via explon
* develop:
  bugfix
  bugfix
  微信调整
  回复规则 bugfix
  后台微信自定义回复重定向bug修复

+++ models/weixin.php
@@ -571,9 +571,18 @@ class weixin_class extends AWS_MODEL
 								$image_file = get_avatar_url($val['question_info']['published_uid'], 'max');
 							}
 							
+							if ($val['associate_action'] == ACTION_LOG::ANSWER_QUESTION OR $val['associate_action'] == ACTION_LOG::ADD_AGREE)
+							{
+								$link = $this->model('openid_weixin')->redirect_url('/m/question/' . $val['question_info']['question_id'] . '?answer_id=' . $val['answer_info']['answer_id'] . '&single=TRUE');
+							}
+							else
+							{
+								$link = $this->model('openid_weixin')->redirect_url('/m/question/' . $val['question_info']['question_id']);
+							}
+							
 							$response_message[] = array(
 								'title' => $val['question_info']['question_content'],
-								'link' => $this->model('openid_weixin')->redirect_url('/m/question/' . $val['question_info']['question_id']),
+								'link' => $link,
 								'image_file' => $image_file
 							);
 						}
@@ -879,8 +888,11 @@ class weixin_class extends AWS_MODEL
 	{
 		if ($reply_rule = $this->get_reply_rule_by_id($id))
 		{
-			unlink(get_setting('upload_dir') . '/weixin/' . $reply_rule['image_file']);
-			unlink(get_setting('upload_dir') . '/weixin/square_' . $reply_rule['image_file']);
+			if ($reply_rule['image_file'])
+			{
+				unlink(get_setting('upload_dir') . '/weixin/' . $reply_rule['image_file']);
+				unlink(get_setting('upload_dir') . '/weixin/square_' . $reply_rule['image_file']);
+			}
 			
 			return $this->delete('weixin_reply_rule', 'id = ' . intval($id));
 		}
